Changed cookbook name
Made the cookbook name all lower case. This prevents berkshelf from complaining:

Berkshelf::BerksError: Berks command Failed: ... reason: In your Berksfile, you have:

  cookbook 'mailcatcher'

But that cookbook is actually named 'MailCatcher'

This can cause potentially unwanted side-effects in the future.

NOTE: If you do not explicitly set the 'name' attribute in the metadata, the name of the directory will be used instead. This is often a cause of confusion for dependency solving.